requirements.txt
# Core dependencies
torch>=2.0.0
numpy>=1.21.0
matplotlib>=3.5.0
scikit-learn>=1.0.0
scipy>=1.7.0
pyyaml>=6.0

# Required by pykan (not auto-installed)
tqdm>=4.60.0
sympy>=1.10.0
pandas>=1.3.0

# KAN implementation
pykan>=0.2.0
